# docs/swagger.yaml
basePath: /
definitions:
...
  category.CategoriesVO:
    description: 获取类目响应
    properties:
      children:
        items:
          $ref: '#/definitions/category.CategoriesVO'
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      path:
        type: string
    type: object
  comment.CommentsVO:
    description: 获取单个评论的响应
    properties:
      account_id:
        type: string
      content:
        type: string
      id:
        type: string
      post_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/comment.CommentsVO'
        type: array
      reply_to_comment_id:
        type: string
    type: object
  dto.CreateCommentRequest:
    properties:
      content:
        maxLength: 1024
        minLength: 1
        type: string
      post_id:
        example: "0"
        type: string
      reply_to_comment_id:
        example: "0"
        type: string
    required:
    - content
    - post_id
    type: object
  dto.CreateOneCategoryRequest:
    properties:
      description:
        type: string
      name:
        minLength: 1
        type: string
      parent_id:
        example: "0"
        type: string
    required:
    - name
    type: object
...
  dto.UpdateOneCategoryRequest:
    properties:
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        example: "0"
        type: string
    required:
    - id
    - name
    type: object
...
  vo.Result:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 响应数据
      msg:
        description: 错误信息
        type: string
      requestId:
        description: 请求ID
      timeStamp:
        description: 响应时间戳
    type: object
host: localhost:9010
info:
  contact: {}
  description: This is the API documentation for Jank Blog.
  version: "1.0"
paths:
...
  /category/createOneCategory:
    post:
      consumes:
      - application/json
      description: 创建新的类目
      parameters:
      - description: 创建类目请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOneCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/category.CategoriesVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 创建类目
      tags:
      - 类目
  /category/deleteOneCategory:
    post:
      consumes:
      - application/json
      description: 根据类目 ID 删除类目
      parameters:
      - description: 类目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/category.CategoriesVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 类目不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 删除类目
      tags:
      - 类目
  /category/getCategoryChildrenTree:
    post:
      consumes:
      - application/json
      description: 根据类目 ID 获取子类目树
      parameters:
      - description: 类目ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/category.CategoriesVO'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 类目不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取子类目树
      tags:
      - 类目
  /category/getCategoryTree:
    get:
      consumes:
      - application/json
      description: 获取类目树
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/category.CategoriesVO'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取类目树
      tags:
      - 类目
  /category/getOneCategory:
    get:
      consumes:
      - application/json
      description: 根据类目 ID 获取单个类目的详细信息
      parameters:
      - description: 类目ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/category.CategoriesVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 类目不存在
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取单个类目详情
      tags:
      - 类目
  /category/updateOneCategory:
    post:
      consumes:
      - application/json
      description: 更新已存在的类目信息
      parameters:
      - description: 类目ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新类目请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOneCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/category.CategoriesVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 类目不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 更新类目
      tags:
      - 类目
  /comment/createOneComment:
    post:
      consumes:
      - application/json
      description: 创建一条新的评论
      parameters:
      - description: 创建评论请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/comment.CommentsVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 创建评论
      tags:
      - 评论
  /comment/deleteOneComment:
    post:
      consumes:
      - application/json
      description: 通过评论 ID 进行软删除
      parameters:
      - description: 评论ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 软删除成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/comment.CommentsVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 软删除评论
      tags:
      - 评论
  /comment/getOneComment:
    get:
      consumes:
      - application/json
      description: 根据文章 ID 获取评论图结构
      parameters:
      - description: 文章ID
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/comment.CommentsVO'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取评论图
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"